// src/types/generated/FeedbackObservabilityReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FeedbackObservability } from "./FeedbackObservability";
import type { FeedbackPersonalizationStatus } from "./FeedbackPersonalizationStatus";

export type FeedbackObservabilityReport = { counters: FeedbackObservability, status: FeedbackPersonalizationStatus, };